MyApps-16: Move web to viewmodel submodule.
